refactor : 검증 관련 변경사항 서비스 레이어에 반영 #44
